Initial commit of Relafix Version 2.0
Dynamic and Static Linkage
==========================
RelaFix now builds an additional shared-object (.so) for dynamic linking.

Data Structures (Arrays, Lists, ...)
====================================
I removed all generic data structures from the project.
If you need an array, use an array[], if you need a dynamic list, use an
intrusive list.  The reason for this decision is that this extra level of
abstraction does not offer any benefit for the project. In fact since few
of these data structures were used in more than one place, their native
counterparts are cleaner and easier to use. On the other hand projects that
use RelaFix will probably be using toolkits like gtk or Qt which offer much
better alternatives.

Intrusive List example
----------------------
Suppose you have a struct like this:
struct foo {
	int id;
	char *name;
};
To make this struct an intrusive list, simply add a pointer to a struct
of the same type.
struct foo {
	int id;
	char *name;
	struct foo *_next;
};
Keep a pointer to the first struct (= the head of the list). The end of the
list is reached if _next is NULL.

Assertions
==========
I removed various "if(foo == NULL) return -1;" blocks.
A procedure should always assume to be used correctly, that is,
it's parameters are non-NULL, indices are positive etc.
If a procedure is called with invalid parameters that is the programmer's
fault and not something that we need to take care of.
Generally speaking, user-input and other things that influence the
programm from the outside should be validated as soon as they are read and
assumed correct from then on.
These assumtions can however be checked with assertions.
Assertions can be used to define pre- and postconditions of procedures aswell
as checking states within a procedure.
Assertions are meant to help the programmer find bugs in his programm and are
automatically removed from the final binary.

Assertions and Unit Tests
-------------------------
Assertions and unit tests serve different purposes. A unit test verifies that
a procedure generates the correct output for a given input, while assertions
merely act as guards to catch obviously invalid programm states.

Errors
======
In RelaFix the structure rf_Error can be used to report errors to the calling
code. The error works like an exception in so far that it contains an
error code and optionally a message and additional data.
If the error code is anything but RF_E_OK the procedures return value should
be considered invalid.
The procedure that uses an rf_Error to report failure must handle the case
that the error pointer passed to it is NULL.
If the error pointer is NULL, the procedure will simply assume no errors occur.
Calling code should use rf_error_reset to free the error message and
additional data.
Note that you don't have to use rf_Error for all error reporting. If there is
only one obvious failure state returning -1 or NULL is fine too.

Usage
-----
rf_Error e;
rf_Relation *r = rf_function_that_can_fail(foo, bar, &e);
if(e.code != RF_E_OK) {
	printf("Failed with code %d: '%s'\n", e.code, e.msg);
	rf_error_reset(&e);
}